# Conflicts:
#   framework-system-view-easyui/src/main/webapp/modules/sys/roleinfo/roleinfo-user.html

+++ framework-system-module/src/main/java/com/centit/framework/system/controller/UserUnitController.java
@@ -130,7 +130,7 @@ public class UserUnitController extends BaseController {
         UserInfo user = sysUserManager.getObjectById(this.getLoginUser(request).getUserCode());
         Map<String, Object> filterMap = convertSearchColumn(request);
         filterMap.put("userCode", userCode);
-        filterMap.put("unitCode", user.getPrimaryUnit());
+//        filterMap.put("unitCode", user.getPrimaryUnit());
 
         List<UserUnit> listObjects = sysUserUnitManager.listObjects(filterMap, pageDesc);
 
@@ -225,6 +225,7 @@ public class UserUnitController extends BaseController {
         UserUnit dbUserUnit = sysUserUnitManager.getObjectById(userunitid);
         if (null == dbUserUnit) {
             JsonResultUtils.writeErrorMessageJson("当前机构中无此用户", response);
+            return;
         }
 
         UserUnit oldValue = new UserUnit();
@@ -233,7 +234,7 @@ public class UserUnitController extends BaseController {
         dbUserUnit.copy(userUnit);
 
         sysUserUnitManager.updateUserUnit(dbUserUnit);
-        
+
         JsonResultUtils.writeSingleDataJson(userUnit, response);
 
         /*********log*********/
@@ -253,6 +254,10 @@ public class UserUnitController extends BaseController {
     public void delete(@PathVariable String userunitid,
                        HttpServletRequest request, HttpServletResponse response) {
         UserUnit dbUserUnit = sysUserUnitManager.getObjectById(userunitid);
+        if("T".equals(dbUserUnit.getIsPrimary())){
+            JsonResultUtils.writeErrorMessageJson("主机构组织信息不能删除！", response);
+            return;
+        }
 
         sysUserUnitManager.deleteObject(dbUserUnit);
